zcash_client_backend: Replace separate `scan_progress` and `recovery_progress` with a single `progress` value.
